requirements.txt
feedparser>=6.0
sqlalchemy>=2.0
pymysql>=1.1
httpx>=0.27
pyyaml>=6.0
apscheduler>=3.10
cryptography>=42.0
trafilatura>=2.0
akshare>=1.14
